+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Clone wrapped component methods on `mediaQueries` HOC.
+- Feature: Add `onCloseClick` prop to `CartRewardCard`.
 - Fix: Update margin on `CartRewardCard` component.
 - Feature: Add `hidePaginationOnMobile` prop to `Carousel`.
 - Fix: Remove `hover` state on `Select-control` class.